+++ backend/src/main/java/com/spring/blog/comment/presentation/CommentController.java
@@ -3,16 +3,16 @@ package com.spring.blog.comment.presentation;
 import com.spring.blog.authentication.domain.Authenticated;
 import com.spring.blog.authentication.domain.user.AppUser;
 import com.spring.blog.comment.application.CommentService;
-import com.spring.blog.comment.application.dto.CommentDeleteRequestDto;
-import com.spring.blog.comment.application.dto.CommentEditRequestDto;
-import com.spring.blog.comment.application.dto.CommentListRequestDto;
-import com.spring.blog.comment.application.dto.CommentListResponseDto;
-import com.spring.blog.comment.application.dto.CommentReplyRequestDto;
-import com.spring.blog.comment.application.dto.CommentResponseDto;
-import com.spring.blog.comment.application.dto.CommentWriteRequestDto;
-import com.spring.blog.comment.presentation.dto.CommentListResponse;
-import com.spring.blog.comment.presentation.dto.CommentResponse;
-import com.spring.blog.comment.presentation.dto.CommentWriteRequest;
+import com.spring.blog.comment.application.dto.request.CommentDeleteRequestDto;
+import com.spring.blog.comment.application.dto.request.CommentEditRequestDto;
+import com.spring.blog.comment.application.dto.request.CommentListRequestDto;
+import com.spring.blog.comment.application.dto.response.CommentListResponseDto;
+import com.spring.blog.comment.application.dto.request.CommentReplyRequestDto;
+import com.spring.blog.comment.application.dto.response.CommentResponseDto;
+import com.spring.blog.comment.application.dto.request.CommentWriteRequestDto;
+import com.spring.blog.comment.presentation.dto.response.CommentListResponse;
+import com.spring.blog.comment.presentation.dto.response.CommentResponse;
+import com.spring.blog.comment.presentation.dto.request.CommentWriteRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
