+++ backend/src/main/java/com/spring/blog/comment/presentation/dto/request/CommentWriteRequest.java
@@ -1,4 +1,4 @@
-package com.spring.blog.comment.presentation.dto;
+package com.spring.blog.comment.presentation.dto.request;
 
 public class CommentWriteRequest {
