+++ backend/src/main/java/com/spring/blog/comment/presentation/dto/response/CommentResponse.java
@@ -1,6 +1,6 @@
-package com.spring.blog.comment.presentation.dto;
+package com.spring.blog.comment.presentation.dto.response;
 
-import com.spring.blog.comment.application.dto.CommentResponseDto;
+import com.spring.blog.comment.application.dto.response.CommentResponseDto;
 import java.time.LocalDateTime;
 
 public class CommentResponse {
