+++ backend/src/main/java/com/spring/blog/post/application/dto/request/PostDeleteRequestDto.java
@@ -1,4 +1,4 @@
-package com.spring.blog.post.application.dto;
+package com.spring.blog.post.application.dto.request;
 
 public class PostDeleteRequestDto {
