+++ backend/src/test/java/com/spring/blog/comment/application/CommentServiceTest.java
@@ -9,13 +9,13 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import com.spring.blog.comment.application.dto.CommentDeleteRequestDto;
-import com.spring.blog.comment.application.dto.CommentEditRequestDto;
-import com.spring.blog.comment.application.dto.CommentListRequestDto;
-import com.spring.blog.comment.application.dto.CommentListResponseDto;
-import com.spring.blog.comment.application.dto.CommentReplyRequestDto;
-import com.spring.blog.comment.application.dto.CommentResponseDto;
-import com.spring.blog.comment.application.dto.CommentWriteRequestDto;
+import com.spring.blog.comment.application.dto.request.CommentDeleteRequestDto;
+import com.spring.blog.comment.application.dto.request.CommentEditRequestDto;
+import com.spring.blog.comment.application.dto.request.CommentListRequestDto;
+import com.spring.blog.comment.application.dto.response.CommentListResponseDto;
+import com.spring.blog.comment.application.dto.request.CommentReplyRequestDto;
+import com.spring.blog.comment.application.dto.response.CommentResponseDto;
+import com.spring.blog.comment.application.dto.request.CommentWriteRequestDto;
 import com.spring.blog.comment.domain.Comment;
 import com.spring.blog.comment.domain.repository.CommentRepository;
 import com.spring.blog.exception.comment.CommentNotFoundException;
