+++ backend/src/test/java/com/spring/blog/comment/presentation/CommentControllerTest.java
@@ -16,16 +16,16 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.spring.blog.authentication.application.OAuthService;
 import com.spring.blog.authentication.domain.user.LoginUser;
 import com.spring.blog.comment.application.CommentService;
-import com.spring.blog.comment.application.dto.CommentDeleteRequestDto;
-import com.spring.blog.comment.application.dto.CommentEditRequestDto;
-import com.spring.blog.comment.application.dto.CommentListRequestDto;
-import com.spring.blog.comment.application.dto.CommentListResponseDto;
-import com.spring.blog.comment.application.dto.CommentReplyRequestDto;
-import com.spring.blog.comment.application.dto.CommentResponseDto;
-import com.spring.blog.comment.application.dto.CommentWriteRequestDto;
+import com.spring.blog.comment.application.dto.request.CommentDeleteRequestDto;
+import com.spring.blog.comment.application.dto.request.CommentEditRequestDto;
+import com.spring.blog.comment.application.dto.request.CommentListRequestDto;
+import com.spring.blog.comment.application.dto.response.CommentListResponseDto;
+import com.spring.blog.comment.application.dto.request.CommentReplyRequestDto;
+import com.spring.blog.comment.application.dto.response.CommentResponseDto;
+import com.spring.blog.comment.application.dto.request.CommentWriteRequestDto;
 import com.spring.blog.comment.domain.Comment;
-import com.spring.blog.comment.presentation.dto.CommentListResponse;
-import com.spring.blog.comment.presentation.dto.CommentWriteRequest;
+import com.spring.blog.comment.presentation.dto.response.CommentListResponse;
+import com.spring.blog.comment.presentation.dto.request.CommentWriteRequest;
 import com.spring.blog.common.PageMaker;
 import com.spring.blog.post.domain.Post;
 import com.spring.blog.user.domain.User;
