+++ backend/src/test/java/com/spring/blog/post/presentation/PostControllerTest.java
@@ -20,14 +20,14 @@ import com.spring.blog.authentication.domain.user.LoginUser;
 import com.spring.blog.common.FileFactory;
 import com.spring.blog.common.PageMaker;
 import com.spring.blog.post.application.PostService;
-import com.spring.blog.post.application.dto.PostDeleteRequestDto;
-import com.spring.blog.post.application.dto.PostListRequestDto;
-import com.spring.blog.post.application.dto.PostListResponseDto;
-import com.spring.blog.post.application.dto.PostResponseDto;
-import com.spring.blog.post.application.dto.PostWriteRequestDto;
+import com.spring.blog.post.application.dto.request.PostDeleteRequestDto;
+import com.spring.blog.post.application.dto.request.PostListRequestDto;
+import com.spring.blog.post.application.dto.response.PostListResponseDto;
+import com.spring.blog.post.application.dto.response.PostResponseDto;
+import com.spring.blog.post.application.dto.request.PostWriteRequestDto;
 import com.spring.blog.post.domain.Post;
-import com.spring.blog.post.presentation.dto.PostListResponse;
-import com.spring.blog.post.presentation.dto.PostResponse;
+import com.spring.blog.post.presentation.dto.response.PostListResponse;
+import com.spring.blog.post.presentation.dto.response.PostResponse;
 import com.spring.blog.user.domain.User;
 import java.util.Arrays;
 import java.util.List;
